Add and delete queue jobs by key
To prevent duplicate jobs and to allow specific jobs to be deleted after they have completed, a combination of variables was used. This commit introduces a change that allows jobs to be added/retrieved with a key instead of a series of variables

This is a necessary change for the URL analyzer.