Fixes for node canvas package on M1 MacBooks
New packages are installed inside the container and a makefile with commands to launch a dev environment has been added

One of the previous issues is that npm commands were ran outside of the container, on the host machine, causing various problems on for different people. The commands in the makefile will run all setup inside the container.